fixed int -> double health conversions

+++ src/main/java/de/raidcraft/skills/DamageManager.java
@@ -117,14 +117,14 @@ public final class DamageManager implements Listener {
         CharacterTemplate character = plugin.getCharacterManager().getCharacter((LivingEntity) event.getEntity());
 
         try {
-            int damage = event.getDamage();
+            double damage = event.getDamage();
             if (environmentalDamage.containsKey(event.getCause())) {
                 if (plugin.getCommonConfig().environment_damage_in_percent) {
                     switch (event.getCause()) {
 
                         case FALL:
                             // the minecraft fall damage is caluclate like so: fall_damage = number of blocks - 3
-                            int height = damage;
+                            double height = damage;
                             damage = (int) (character.getMaxHealth() * (environmentalDamage.get(event.getCause()) * height));
                             break;
                         case BLOCK_EXPLOSION:
@@ -134,7 +134,7 @@ public final class DamageManager implements Listener {
                             // default damage is as follows: 97 (charged creeper), 65 (TNT), 49 (creepers), 17 (fireballs)
                             // (1 x 1 + 1) defines the range the entity is away and how much the power gets reduced
                             // damage = (1 × 1 + 1) × 8 × power + 1
-                            float power = (damage - 1) / 8;
+                            float power = (float) ((damage - 1) / 8);
                             damage = (int) (character.getMaxHealth() * (environmentalDamage.get(event.getCause()) * power));
                             break;
                         default:
@@ -147,7 +147,7 @@ public final class DamageManager implements Listener {
             }
             event.setCancelled(true);
             if (damage > 0) {
-                EnvironmentAttack attack = new EnvironmentAttack(event, damage);
+                EnvironmentAttack attack = new EnvironmentAttack(event, (int) damage);
                 attack.run();
             }
         } catch (CombatException e) {
